# Improve performance of BigInteger.Multiply