--- a/src/System.Runtime.Numerics/tests/BigInteger/op_multiply.cs
+++ b/src/System.Runtime.Numerics/tests/BigInteger/op_multiply.cs
@@ -18,6 +18,13 @@ namespace System.Numerics.Tests
             byte[] tempByteArray1 = new byte[0];
             byte[] tempByteArray2 = new byte[0];
 
+            // Multiply Method - One Large BigInteger
+            for (int i = 0; i < s_samples; i++)
+            {
+                tempByteArray1 = GetRandomByteArray(s_random);
+                VerifyMultiplyString(Print(tempByteArray1) + "u*");
+            }
+
             // Multiply Method - Two Large BigIntegers
             for (int i = 0; i < s_samples; i++)
             {
@@ -158,6 +165,13 @@ namespace System.Numerics.Tests
         {
             byte[] tempByteArray1 = new byte[0];
             byte[] tempByteArray2 = new byte[0];
+
+            // Multiply Method - One Large BigInteger
+            for (int i = 0; i < s_samples; i++)
+            {
+                tempByteArray1 = GetRandomByteArray(s_random);
+                VerifyMultiplyString(Print(tempByteArray1) + "u*");
+            }
 
             // Multiply Method - Two Large BigIntegers
             for (int i = 0; i < s_samples; i++)