To introduce performance tweaks for BigInteger's operations a new static class `BigCalc` implements those algorithms based on raw uint[] objects, partially switching even to pointer arithmetic to spare some index calculations. First of all a multiplication implementation is included, furthermore an optimized version for squaring.

A basic performance comparison based on [this code](https://gist.github.com/axelheer/3b701c91271b7c762586) unveils the following results:

**Multiply**

| # of bits | # of vals | before ms | after ms |
|----------:|----------:|----------:|---------:|
|        16 | 1,000,000 |        41 |       35 |
|        64 | 1,000,000 |       127 |       89 |
|       256 | 1,000,000 |       284 |      182 |
|     1,024 |   100,000 |       236 |      147 |
|     4,096 |    10,000 |       323 |      142 |
|    16,384 |     1,000 |       495 |      132 |
|    65,536 |       100 |       785 |      121 |

**Square**

| # of bits | # of vals | before ms | after ms |
|----------:|----------:|----------:|---------:|
|        16 | 1,000,000 |        41 |       42 |
|        64 | 1,000,000 |       127 |       78 |
|       256 | 1,000,000 |       284 |      147 |
|     1,024 |   100,000 |       236 |      102 |
|     4,096 |    10,000 |       323 |       98 |
|    16,384 |     1,000 |       495 |       92 |
|    65,536 |       100 |       785 |       84 |

*Note:* the performance gain is not as big as expected based on my previous work (as mentioned within the issue), but I'd still call it significant. I actually had some difficulties executing a custom build of *System.Runtime.Numerics* though, since just dropping in the *dll* as described [here](https://github.com/dotnet/corefx/wiki/Open-Source-Signing) does not work 'cause of some "type forwards". I finally managed to execute it using *kre-coreclr-win-x64.1.0.0-beta3* and some hacks... :flushed:

#1307