@ellismg 2015-04-20
@dotnet-bot test this please