@KrzysztofCwalina 2015-04-21
BTW, the perf results are impressive. Thanks for doing this PR! I think it's a great example how the community members with domain expertise, like yourself, can make .NET better. 

I added some comments to the PR. The main one is that I wish the sources had more comments and some links to the algorithm description.