@sharwell 2015-04-22
@hpshelton The cases you are pointing out regarding curly braces already conform to the style guidelines for this project.