@hpshelton 2015-04-22
Thanks, @sharwell. Deleted all but one instance of the comment.