@axelheer 2015-04-22
Thanks! I'll update the code accordingly.