@stephentoub 2015-04-24
@dotnet-bot test this please