@stephentoub 2015-04-26
@dotnet-bot test this please