@KrzysztofCwalina 2015-04-27
This looks good to me. Please squash the commits and I will merge this PR.