@KrzysztofCwalina 2015-04-28
Thanks for the contribution. It's a very nice improvement!